# swagger.yaml
openapi: 3.0.0
info:
    title: 'Ecommerce Application'
    version: 1.0.0
    description: 'API documentation for Ecommerce application'
components:
    securitySchemes: {Auth: {type: apiKey, in: header, name: Authorization}}
security:
    - {Auth: []}
basePath: /
paths:
    /api/v1/auth/users: {get: {summary: 'Get logged-in user data', tags: [Authentication], security: [{BearerAuth: []}], responses: {'200': {description: 'Successful response', content: {application/json: {example: {_id: 603ddca3fd6e1d396055bee0, name: 'John Doe', email: john@example.com, createdAt: '2022-02-01T10:00:00.000Z', updatedAt: '2022-02-01T12:00:00.000Z'}}}}, '401': {description: Unauthorized}, '500': {description: 'Internal server error'}}}}
    /api/v1/products: {post: {summary: 'Get all products', tags: [Products], security: [{Auth: []}], requestBody: {required: false, content: {application/json: {example: {category: electronics}}}}, responses: {'200': {description: 'Successful response', content: {application/json: {example: {category: electronics}}}}, '401': {description: Unauthorized}, '500': {description: 'Internal server error'}}}}
    /api/v1/products/create: {post: {summary: 'Create a new Product', tags: [Products], security: [{Auth: []}], requestBody: {required: true, content: {application/json: {example: {text: 'What is your question?', options: [{id: 1, text: 'Option 1'}, {id: 2, text: 'Option 2'}], correctOption: [1], courseType: Math}}}}, responses: {'200': {description: 'Successful response', content: {application/json: {example: {msg: 'Product inserted Successfully!'}}}}, '400': {description: 'Bad Request'}, '500': {description: 'Internal server error'}}}}
    /api/v1/users/signup: {post: {summary: 'User SignUp', tags: [Users], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {username: {type: string, description: 'username should be unique'}, email: {type: string, format: email, description: 'User''s email address'}, password: {type: string, format: password, description: 'User''s password (at least 6 characters)'}}}}}}, responses: {'200': {description: 'Successful signup', content: {application/json: {example: {token: '<JWT_TOKEN>'}}}}, '400': {description: 'Bad request (validation errors)', content: {application/json: {example: {errors: [{msg: 'Please enter valid username.'}, {msg: 'Please enter your email address.'}, {msg: 'Password must be strong (include uppercase, lowercase, number, and special character).'}]}}}}, '409': {description: 'Conflict (user already exists)', content: {application/json: {example: {msg: 'User already exists'}}}}, '500': {description: 'Internal server error', content: {application/json: {example: {msg: 'Server error'}}}}}}}
    /api/v1/users/login: {post: {summary: 'Vendor login', tags: [Users], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {email: {type: string, format: email, description: 'User''s email address'}, password: {type: string, format: password, description: 'User''s password (at least 6 characters)'}}}}}}, responses: {'200': {description: 'Successful login', content: {application/json: {example: {token: '<JWT_TOKEN>'}}}}, '400': {description: 'Bad request (validation errors)', content: {application/json: {example: {errors: [{msg: 'Please enter your email address.'}, {msg: 'Please insert at least 6 characters.'}]}}}}, '409': {description: 'Bad request ( user not exists)', content: {application/json: {example: {msg: 'User Not exists'}}}}, '500': {description: 'Internal server error', content: {application/json: {example: {msg: 'Server error'}}}}}}}
    /api/v1/users/vendor-signup: {post: {summary: 'Vendor SignUp', tags: [Vendors], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {username: {type: string, description: 'username should be unique'}, email: {type: string, format: email, description: 'User''s email address'}, password: {type: string, format: password, description: 'User''s password (at least 6 characters)'}}}}}}, responses: {'200': {description: 'Successful signup', content: {application/json: {example: {token: '<JWT_TOKEN>'}}}}, '400': {description: 'Bad request (validation errors)', content: {application/json: {example: {errors: [{msg: 'Please enter valid username.'}, {msg: 'Please enter your email address.'}, {msg: 'Password must be strong (include uppercase, lowercase, number, and special character).'}]}}}}, '409': {description: 'Conflict (user already exists)', content: {application/json: {example: {msg: 'User already exists'}}}}, '500': {description: 'Internal server error', content: {application/json: {example: {msg: 'Server error'}}}}}}}
    /api/v1/users/vendor-login: {post: {summary: 'Vendor login', tags: [Vendors], requestBody: {required: true, content: {application/json: {schema: {type: object, properties: {email: {type: string, format: email, description: 'User''s email address'}, password: {type: string, format: password, description: 'User''s password (at least 6 characters)'}}}}}}, responses: {'200': {description: 'Successful login', content: {application/json: {example: {token: '<JWT_TOKEN>'}}}}, '400': {description: 'Bad request (validation errors)', content: {application/json: {example: {errors: [{msg: 'Please enter your email address.'}, {msg: 'Please insert at least 6 characters.'}]}}}}, '409': {description: 'Bad request ( user not exists)', content: {application/json: {example: {msg: 'User Not exists'}}}}, '500': {description: 'Internal server error', content: {application/json: {example: {msg: 'Server error'}}}}}}}
tags:
    - {name: Authentication, description: 'User authentication and profile retrieval'}
    - {name: Products, description: 'API operations related to products'}
    - {name: Users, description: 'User Signup and Login'}
    - {name: Vendors, description: 'Vendor Signup and Login'}